+++ src/colors/ColorConverterHSV.h
@@ -8,6 +8,6 @@ public:
 	ColorRGB TransformToRGB(Color p_currentColor) override;
 	Color TransformFromRGB(ColorRGB p_currentColor) override;
 	ofFloatColor GetColor(Color p_currentColor) override;
-	std::vector<ofParameter<float>> getParameters(Color p_currentColor) override;
+	std::vector<ofParameter<int>> getParameters(Color p_currentColor) override;
 };
